Changes to both the CSS and Index.HTML Files
[ADDED]
Added a H1 Title tagged, "Minecraft Bingo"
Added a structure the website, titled Top Mid and Bottom Sections
Added Stying for the Generate new card button and renamed it to "New
Card"
Added a Rules and Questions Section in the bottom section. [Needs
Detailing]
[Changes]
Changed the width and hight of the table rows boxes to a bit smaller
than the default values.
Changed settings layout and added some padding and a background color.
Changed the border width of the table boxes.