Difficulty setting, goal frequency and more
Added a difficulty setting, from 1 to 5. The setting changes the amount
of medium/hard/veryhard goals on the sheet, 1 being all simple goals and
5 being almost none. Added a slider for the setting, complete with css
styling, and also updated the url generation.

Added a goal frequency type, which allows goals to have a custom chance
to show up on the board, which is currently used for repeat goals such
as potions which have 9 entries to show up 9 times as less (meaning it's
the same chance as any other single goal).

Dissabled the layout button and removed it from the default url.

Renamed and moved options dropdown javascript code, as well as renaming
the functions/ids to be less "clearly copy pasted from the internet
CHANGEME" names.

Tweaked a few goals, added a few new ones.