Add versioning, some refactoring
The bingo.js now contains a list of versions, each referring to a list of goals and a generator function, which are stored in separate .js files. The generator function is responsible for selecting and returning the goals which should appear on the sheet, which before was integreated into the generateNewSheet() function and is now separated to allow for different algorithms per version (in addition to goals).

Also removed some global variables that weren't really necessary (the generator function shouldn't access any global variables) and did some other refactoring to improve readability.

Currently version 1 is intended to be the first stable version (mostly for testing), so the associated files shouldn't be changed anymore. For new dev stuff the "dev" version can be used with the "v2" files. Those files are then intended to be used for the actual version 2 once that is ready (would also be possible to do it the other way around, have dedicated "dev" files for the "dev" version which will then be copied over to new stable version files). Of course you wouldn't want to release a new version for every change, but much more rarely, when it's apparent that not a lot new changes will happen/have happened or something.

Also added a section for the user to select a version and a notice when the user isn't using the latest stable version. If the user comes to the page without a version in the URL it will automatically go to whatever is defined as the latest stable version.