Moved Tame an Ocelot, added Jungle Wood
Finding a Jungle isn't the only challenging part about taming an Ocelot,
as such it was too difficult for a Medium goal so it was moved to Hard.
Added Jungle Wood as a replacement in Medium to fill the "find a jungle"
type goal. It keeps the frequency of 5 to stop there being too many Wood
goals when playing on Medium (because of all the Easy Wood goals)