Added Tag system
Added a Tag system to the bingo goals and generation to address the issues with certain types of goals being overbearing on the sheet.

Now each goal can (and should) have at least one tag attached, which is a basic description of the goal. Tags have a name and a maximum amount. If the maximum amount of tags is already present on the bingo sheet no more will be added. EG: There can only be 2 "RareBiome" tagged goals at any time on the sheet.