use URLSearchParams to parse query part of the URL
Create an instance of URLSearchParams and use its method get() to parse
parameters in query part of the URL in the function gup(), instead of
manually parsing it with RegExp.  Link to the documentation of API in
the comment above the function.  Reformat the function for consistency
with the rest of the file, while we're here.

This interface is supported by all major browsers except Internet
Explorer.[1] However, use of some other [2] ES6 features already makes
Bingo incompatible with IE.

Footnotes
1. https://www.caniuse.com/urlsearchparams
2. Like destructuring assignment in function shuffle, see
   https://stackoverflow.com/a/50042740/1083697