add version selection dropdown menu
Intoduce a new button: versions button.  Versions button toggles a new
menu, which is populated by function fillVersionSelection().  Individual
version buttons are similar to existing options buttons, except that
they are 50px wider to accommodate the length of versions' label texts.
To make it fit, the label for 'dev' version is made shorter by replacing
'in development' with 'WIP' (abbreviation of "work in progress").