css: fix vertical misalignment of text on buttons
Only button .new-seed-button has 'line-height: 50px', while all other
buttons do not.  So drop this attribute from .new-seed-button.  Adjust
the line-height for text in sliders, so that their text is aligned with
the text on the buttons.  Note, that .slider-text needs the line-height
attribute to center the text vertically---without it, the text is
aligned to top by default.

Text in tags <button> is centered vertically by default.  I tried using
flexbox [1] to center text in sliders vertically, but HTML and CSS
become way too complicated---just one hardcoded line-height is better.

[1] https://stackoverflow.com/a/38543056/1083697