css: justify buttons using flexbox
In future commits, more buttons will be added to .buttons-row.  To
prepare for it, replace individual 'margin-right: 8px' of class
.new-seed-buttons with a flexbox container.  With new flex rules, the
leftmost button in .buttons-row will be always aligned with the left
border of .buttons-row and the rightmost button with the right border.

This is the first time I've used flexbox CSS after learning it using
this tutorial: https://www.tutorialspoint.com/flexbox/index.htm