js: use hide() & show() in updateStreamerMode()
Function updateStreamerMode() shows and hides individual elements by
setting appropriate value for the CSS attribute 'display'.  Extract
selectors for these elements into a constant and bulk hide() or show()
them.  Note, that when hide() is used, jQuery remembers the current
value of the 'display' attribute to be reinstated when show() is called.