js: make "Exit Streamer Mode" smoother
When Streamer Mode is enabled using the button in the "Options..." menu,
the dropdown of the menu is left as is.  When "Exit Streamer Mode" is
clicked, the dropdown is hidden with an animation by the dropdown menu
handling function, because it was a click outside of a dropdown.  This
animation looks like a glitch, compared to all other elements, which are
shown/hidden immediately on Streamer Mode switching.  Since we are
planning to add more dropdown menus in the future, which might make this
animation more jarring.

So hide the dropdown immediately whenever Streamer Mode is toggled, so
that "glitchy" animation is not shown.  We can do it safely, since
Streamer Mode itself doesn't have any dropdown menus and users don't
expect to see the dropdowns after exiting Streamer Mode.