js: hide the whole #bingo on "Hide Table" click
Function updateHidden() hides individual table cells ('#bingo td') to
hide the contents of the bingo sheet when "Hide Table" button is
clicked.  It also goes through all tootlip icons in the #bingo table to
show them when the table is shown (HIDDEN is false-y).  This is done to
change from their default CSS 'visibility: hidden'.  The #bingo table
itself, however, stays visible, showing its background (stone-tile.png).
This background is covered by table cells when the sheet is shown.

So overhaul how the sheet hiding is done to simplify updateHidden().  To
allow hiding the whole #bingo table, move most of its styling (border,
width, etc) to a new wrapper div#bingo-box.  The background image,
however, is moved to a new div#hidden-bingo, which is wrapped around
#hidden-table.  This wrapping allows us to simplify the CSS for position
and size of #hidden-table.  See declaration 'padding: 50px' for
#hidden-bingo.  This makes the width of the stone-textured border around
#hidden-table consistently 50px.  Judging by the old values of width and
margin-top, this was probably the intended positioning of #hidden-table.

The wrapper #hidden-bingo is also placed inside the new #bingo-box, so
it's a sibling of #bingo.  Both #bingo and #hidden-bingo are set to fill
out the inside of #bingo-box (see height and width declarations).

Hide the whole #bingo table and show #hidden-bingo when "Hide Table" is
clicked using new CSS class "hidden", which is added or removed from the
container #bingo-box in the function updateHidden().