hide .tooltipQ using CSS attribute rule
Hiding and showing of the tooltip icon .tooltipQ was broken by the
recent commit fa40258 (js: hide the whole #bingo on "Hide Table" click,
2021-03-06), because it removed a bit of code which touched visibility
of tooltips.

Steps to reproduce:

  1. Open a URL with a hidden table setting, for example
     index.html?s=4-1-0-dev_89001
  2. Click "Show Table"
  3. Observe the squares which are supposed to have tooltips,
     e.g. "Activate a 5x4 Nether Portal (not counting corners)".

So overhaul the .tooltipQ showing/hiding code simplify it.  This also
fixes the bug.  Set the data attributes to individual squares using the
attr() setter.  We have to use `attr()` setter, rather than `data()`,
because, quote from jQuery documentation [1]:

> Using the `data()` method to update data does not affect attributes
> in the DOM. To set a `data-*` attribute value, use `attr`.

So use attr() to set data attributes of squares in generateNewSheet(),
thus making it available to new CSS selectors.  Hide the .tooltipQ using
CSS, when the data attribute for the tooltip text in the <td> tags
(bingo squares) is empty.

[1] https://api.jquery.com/data/ see also
    https://stackoverflow.com/a/13524139/1083697