bingo.js: update #tooltip when hovering over square
Currently, the contents of #tooltip are filled in when hovering over
"#bingo td img", i.e. the tooltip icon.  Because of that, user may see
the old contents of the tooltip for a split moment before it had been
updated.

Fill in the contents of #tooltip every time user hovers over a square to
fix this visual glitch.