bingo.js: don't check attributes when showing the tooltip
Element "#bingo td img" is hidden using 'display: block', when there is
no tooltip text for that square.[1]  That means that it is not possible
to hover over <img> in a square without tooltip.  So stop checking the
attributes of the square, which is the parent of <img> when showing the
tooltip, since we know that the condition in the `if` is always false.

[1] see CSS rule for .tooltipQ introduced in commit 48ce64b (hide
    .tooltipQ using CSS attribute rule, 2021-03-07)