bingo.js: store COLOURCOUNT in localStorage
Changing colour count to user preference every time is annoying.  So
help the user by storing and loading the COLOURCOUNT setting to and from
browser's local storage.[1]  The try-catch around the call to
localStorage.setItem is needed, because Mobile Safari will throw an
exception in private mode.[2]

Before, function updateColourCount() on page load only set the slider
text to the default value.  Now that function updateColourCount() may be
dealing with non-default values coming from localStorage, update the
state of the slider as well as text.

Note, that most browsers don't save localStorage when using 'file://'
protocol.[3]  A local HTTP server [4] can be used for testing.

[1] https://developer.mozilla.org/en-US/docs/Web/API/Window/localStorage
[2] https://developer.mozilla.org/en-US/docs/Web/API/Storage/setItem
[3] https://stackoverflow.com/a/24997489/1083697
[4] for example, the builtin HTTP server in Python
    https://docs.python.org/3/library/http.server.html