bingo.js: save COLOURCOUNT in cookies
Changing colour count to user preference every time is annoying.  So
help the user by storing and loading the COLOURCOUNT setting to and from
browser cookies.  This is the first time I used cookies from JavaScript
and I used tutorials [1] and [2].  The `SameSite` part of the cookie is
explained in [3].

Since this a user preference, not related to any tracking or
advertisement, this cookie should not require consent.[4]

Footnotes
1. https://developer.mozilla.org/en-US/docs/web/api/document/cookie
2. https://www.w3schools.com/js/js_cookies.asp
3. https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Set-Cookie/SameSite
4. https://www.privacypolicies.com/blog/eu-cookie-law/#Some_Cookies_Don_T_Need_Consent