html: use only 'oninput' event for slider updates
Because both 'oninput' and 'onchange' events are used for updates of
both sliders (colour count and difficulty), the functions are called
twice: first time slider thumb is moved ('oninput' event), and a second
time on mouse up ('onchange' event).

So remove 'onchange' event handlers, so that these function are only
called once.