Goal Tweaks
Removed "Gold Block" from Very easy and adjusted the other difficulties.
Changed "(2-5) Apples" to Easy and added "Apple" to Very Easy
Added RareBiome tag to Pillager goals as they basically require a Pillager Outpost and show up on sheets too often
Moved "Grow a Huge Fungus" to Hard and added the "Nether" to the name to avoid confusion with Mushrooms.