bingo.js: make "6 Colours" consistent with shorcuts
Default option for COLOURCOUNT is "Blue, Green, Red".
Current order for shorcuts is:

  0 no colour (grey)
  1 blue
  2 red
  3 green
  4 yellow
  5 pink
  6 brown

So change the order of option "6 Colours" to be consistent with the
order of colours set by shortcuts.  That way it is also consistent with
the default "Blue, Green, Red" and additional three colours are "added"
to the end.