bingo.js: highlight goals with "Never" tag
Goals with tag "Never" are always completed at the very start at the
game.  Playtesting has shown that players sometimes don't notice the
"Never" goals and lock themselves out of them before reading the sheet
completely.

Most people use green to signify completion of a goal.  So highlight all
the "Never" goals as if they are already completed when the sheet is
generated.  Another argument for the green colour is that it is the
colour available in all colour count options.