generator_v3: don't check all previously generated goals unnecessarily
When iterating over previously generated goals, generator_v3 checks
different conditions, which might prevent a goal candidate to be chosen:
frequency check, same goal on sheet, antisynergry, catalyst/reactant,
and tag counting.  This is tracked using a boolean variable `cont`.  It
signifies, whether or not the currently chosen goal candidate satisfies
those conditions.  It is checked in a `do {...} while (!cont);` loop.

Whenever one of the conditions for a good goal is failed, `cont` is set
to `false`.  At this point, we've decided that `goalCandidate` is no
good, so we can skip the other checks.  Note, that this refactoring
preserves bingo sheet generations as is.

While this optimization most certainly won't be noticeable by users, it
is noticeable when running tests.  For example, when calculating
statistics for sheets for all seeds and all difficulties, this optimized
version runs 20-25% faster on my laptop, cutting the running time to
just under five minutes.