Goal changes
Added #73
Fixed #75
Changed "5-15 Redstone" to "Redstone" as this allows more options for finding it outside of mining.
Removed "Build a Redstone AND Gate" because it's bad for loads of reasons
Changed "Activate a Target Block" to "Open a Door with a Target Block from 10 blocks away" to be clearer about the goal, and provide an easy visual way to know you did it.
Changed all Bed Colour goals into Placing Bed goals
Gave Bounce on a Bed an Overworld tag