bingo.js: right click to switch to previous square colour
Currently, left mouse button click on a square switches its colour to
the next colour in the colour selection.  Switch bingo squares to the
previous colour, when the user clicks with right mouse button on a
square.

In function `anotherColour`, adding `colourSelection.length` to
`currIndex + increment` is needed, because we want the left operand of
`%` to be non-negative.