css: use generic .tooltip instead of specific #tooltip
In the future commits, we might want to add more tooltips to the
website.  To make styling of tooltips easier, turn most of the rules for
id '#tooltip' into rules for class '.tooltip'.  Rename id '#tooltip'
into '#goalTooltip' to make the currently used tooltip more distinct.

Note, that some CSS rules are specific to #goalTooltip, because goal's
tooltip has an image on the left, which requires moving the <span> a
little bit.