options menu: use style of Minecraft pause menu
In the future, more options will be added (e.g. Dark Mode).  The
dropdown menu with all options is getting too big to be convenient to
use.  So move the controls for options into a new menu, which opens over
the bingo sheet in the style, similar to Minecraft's pause menu.

- Open dialog on the click of "Options..." button
- Close dialog when:
  - clicking outside of the dialog controls
  - clicking on the special "x" button in the dialog (top right corner)
- Toggle the dialog when the Esc key is pressed

In the process, we also reorder the options, so that the most important
options are higher in the menu.

Showing the options menu in the style of Minecraft pause menu was
originally Wayno's idea:
https://discord.com/channels/131720430326120449/819186365085253673/823334300644671498