Refactored generator_v3
Removed old comments, added new comments
Removed unused variables
Better creation of Indexes array
Added tagCount list which keeps track of tags on the sheet and their amounts instead of recalculating the list/amount by checking every goal on the sheet every time a goal is generated!
Added antisynergy, reactant and catalyst arrays, to keep track of what antisyergys etc are on the sheet instead of checking every goal on the sheet every time a goal is generated!
Refactored check for if the goal is already on the sheet.
Refactored check for same line tag collisions.