cont -> retry
Renamed cont to retry, moved it outside the do while loop (how did this even work anyway?).
Removed constant refrences to cont when failing a goal check and instead just set retry to false ONCE at the end of the do while loop is all the checks are passed.
Named do while loop GoalGen and continue that specifically when failing a check.