generator_v3: use Sets to track antisynergys, reactants, and catalysts
generator_v3 keeps track of antisynergys, reactants, and catalysts on
the sheet using three arrays.  Logic for all three of those requires a
unique set of strings, and two of these arrays (reactants and catalysts)
are deduplicated for every generated goal.

So use the Set type for keeping track of antisynergys, reactants, and
catalysts of goals, already added to the sheet.

  https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Set

Fix the spelling in comments, while we're here.