goals_v4: add tag Overworld to Bookshelves goal
Bookshelves can be either crafted or mined with Silk Touch from
generated structures.  Crafting recipe includes any Planks (including
Huge Fungus planks) and Books.

The only way to obtain Books without crafting them in the Nether is to
disenchant Enchanted Books (Soul Speed books can be found in the Bastion
remnant Chests).  However, this requires a Grindstone, which can not be
obtained in the Nether or the End, because it can only be crafted using
a Stone Slab.  There is no way to obtain Paper to craft Books in the
Nether or the End.

So add tag Overworld to the "Bookshelves" goal.