options menu: blur background when menu is shown
"Pause menu"-style options menu looks bad when the table is hidden,
because "Hide Table" mode has buttons with lots of text around it, and
it is hard to look at options menu controls overlaid over controls of
the "Hide Table" mode.  So blur the background when the options menu is
shown, so that controls in the options menu are easier to make out.