move version dropdown to options menu from .buttons-row
Elements in the #info_bar are ordered by what you change the most:

   1. Seed (most often)
   2. Difficulty
   3. Version (least often)

However, buttons in the .buttons-row are ordered:

   1. Version
   2. Seed
   3. Difficulty
   4. Options

Move the version dropdown to the recently redesigned options menu. Make
the toggle button of the dropdown wider and use the new width to show
the Minecraft version on the button in addition to the Bingo version.