Change Preload to Prefetch, add two more images
Preload will force the browser to download the images on load, thus slowing down the loading of the page. Prefetch is what we intended to use, which only downloads once the browser is idle.
Added the slider background and the hide sheet images to pretetch as both of these suffer from the same problems the symbols did.