js: use jQuery to calculate position of Copy Seed buttons
When a user clicks on a "Copy Seed" button, function showCopiedTooltip()
uses event.target's offsets to calculate the position at which the
"Copied!" tooltip will be shown.  In commit [1], `position: relative`
has been added to #bingo-box, so that .pause-menu could have `position:
absolute`.  This broke calculation of the position for the tooltip shown
near the "Copy Seed" button in the "Hide Table" mode.

Use jQuery's offset() function to calculate absolute coordinates for the
tooltip correctly, regardless of the CSS attribute `position`.

[1] 0275bbb (options menu: use style of Minecraft pause menu,
    2021-03-19)