use CSS for showing export
When refreshing index.html, the popup window for JSON export flashes for
a split second, because it is shown by default and is hidden by
JavaScript after the page is rendered.

Set new CSS class .show-export to <body> in functions showGoalExport()
and hideGoalExport() instead of showing/hiding #export itself.  Since
export dialog would be hidden by default via CSS, there is no need to
hide it on load via JS.

This is similar to the refactoring done in commit 9fe92f2 (use CSS for
Streamer Mode, 2021-03-07).