hide tooltips by default using CSS
Hide the tooltips (goal hints and "Copied!" popup) by default using a
CSS declaration instead of hiding them in bingo.js.  Showing the
tooltips by default doesn't cause any visual glitches -- they are
positioned at the bottom of the screen before JavaScript starts
execution, so they are always out of the view when the .hide()ing
happens.  Doing this in CSS, however, is less code, which is nice.