css: remove unnecessary styling of tag <html>
Firefox has interesting behaviour with regard to the colour of the
scroll bar.  It determines the colour of the scrollbar from
background-color of tags <html> and <body>. Because CSS class .dark is
added to <body>, it means that tag <html> stays with the light mode
value of `--bg-colour` variable.  This causes the scroll bar to conform
to light mode when dark mode is enabled.

So remove CSS rules for <html>, since we already have rules for <body>.
There are some differences between styling of <html> and <body> [1], but
we don't hit those because of `min-height: 100%` rule for <body>.  The
JS code in bingo.js for background image randomization, streamer mode,
and dark mode already uses <body> tag.

An alternative to removing styles for <html> would be to switch
everything else from using tag <body> to use <html> instead.

[1]: https://css-tricks.com/just-one-of-those-weird-things-about-css-background-on-body/