goals_v5: add Overworld tag to potions
Only "Potion of Fire Resistance" can be obtained without visiting
Overworld -- via bartering with Piglins.  All the other potion goals
require Water Bottles, which can only be obtained in the Overworld, as
far as I can tell.[1]  There are also ways of getting some potions
without brewing -- Buried Treasure, Witches, Ancient City [2] -- but all
these are limited to Overworld too.

Goals for Potion of Regeneration and Potion of Strength are missing the
Overworld tag.  It seems to me that when all other potion goals got the
Overworld tag added to them in commit [3], only the brewing ingredients
were considered.  The three potion goals which are missing the tag are
all brewed with Nether ingredients:

  - Blaze Powder for Potion of Strength
  - Ghast Tear for Potion of Regeneration
  - Magma Cream for Potion of Fire Resistance (though also available
    through bartering)

Add tag `Overworld` to two potion goals which are missing it in
goals_v5.js.[4]  Update comment near the goal for Potion of Fire
Resistance correspondingly.  See also the discussion in Discord.[5]

[1] https://minecraft.wiki/w/Potion#Obtaining
[2] only if you're playing Bingo on the wrong version of Minecraft ;-)
[3] 7a6fb99 (Variable Tag Maximums by Difficulty, 2021-03-01)
[4] Remember, goals_v4 and earlier are fixed in stone now.
[5] https://discord.com/channels/131720430326120449/1045020502621691985/1160398977141846026