Add TheRun.gg Race Listing Support (#138)
* TheRun.gg Races Display

Replaced the old "Playing with Others" section with a new "TheRun.gg Races" section.
Upon pressing the load races button, it fetches all the current Minecraft: Java Edition races from TheRun.gg (via a cloudflare worker I setup) and then filters the Bingo races into a table list.
You can then click on a race in that list to open the race on TheRun.gg in a new tab.
If the description of the race contains a MinecraftBingo.com link, it will also automatically update the page to that link.

* Added race seed copying support

Now when a TheRun.gg race has a sheet listed, clicking it will automatically copy the seed to your clipboard as well as loading the sheet.

* Updated TheRun.gg URLs

The tip prompting you to create your own race now links you directly to the race creation page.
All TheRun.gg links now open in a new tab.