File Clean up (#139)
Cleaned up the general file structure.

Added folders for Images, Fonts, Generators and Goals. Placed the appropriate files in each folder.

Updated (hopefully) all file paths related to the moved files.

Added the tooltip folder structure to bingo.js and removed the folder structure from each individual goal, so now the goals only need to list a file name. If the goal uses an external image for the tooltip the leading folder structure isn't not used.

Moved the Question Mark image to the Symbols folder along with the other Symbols.