Added Popout button in Options
Added a dedicated popout button in the options menu. Previously it was only possible to popout in streamer mode.

Now if the page is popped out and it isn't in streamer mode it scrolls down the correct amount to display the bingo sheet neatly in the popout window.

Added a few more optional parameters to the popout window open function which may or may not be respected by certain browsers.