Added Transition Toggle to Dark Mode
Added a transition toggle to updateDarkMode.

Now when the page first loads and finds a user setting for dark/light mode it will not play the transition and instead just go straight to the user's preference, preventing the screen from flashing.

Pressing the toggle dark mode button will still play the transition.