Reduce loading flash
Set the body of the page to "dark" in the HTML so that when the page initially loads it defaults to the dark theme to stop the screen flashing while loading.

The user's setting of dark vs light more is still respected, and the page still loads light mode by default.

Had to change line 90 of bingo.js to fix the backgrounds not displaying with this change.