Initial Custom Tag Maximum Support
Custom tags maximum can be defined in the URL as such:
&tags=TAGNAME_MAXAMOUNT-SECONDNAME_MAXAMOUNT etc

This is only supported in the latest generation version.

The generator now keeps an array of the tags objects on the sheet as opposed to just a count of how many of each tag is on the sheet.

Added basic debug console logging of the tags on the sheet.