Tag customisation box improvements
If a tag was not changed from the default, it now updates when the difficulty is changed to the new default.
Refactored the slider updating to a function instead of lots of different places/things.
When a tag is changed from the default it now changes text colour to red.